requirements.txt
keystoneauth1
oslo.serialization
grequests